+++ HISTORY.rst
@@ -44,6 +44,7 @@ Bug fixes
 * Fix bug in ``missing_wmo``, where a period would be considered valid if all months met WMO criteria, but complete months in a year were missing. Now if any month does not meet criteria or is absent, the period will be considered missing.
 * Fix bootstrapping with dask arrays. Dask does not support using ``loc`` with multiple indexes to set new values so a workaround was necessary.
 * Fix bootstrapping when the bootstrapped year must be converted to a 366_day calendar.
+* Virtual modules and translations now use 'UTF-8' by default when reading yaml or json file, instead of a machine-dependent encoding.
 
 Internal Changes
 ~~~~~~~~~~~~~~~~